Added data to the database
