+++ database/data/difficulty.csv
@@ -1,1 +1,40 @@
-3,Easy
+97,360DegreeEasy
+100,360DegreeExpert
+101,360DegreeExpertPlus
+99,360DegreeHard
+98,360DegreeNormal
+102,90DegreeEasy
+105,90DegreeExpert
+106,90DegreeExpertPlus
+104,90DegreeHard
+103,90DegreeNormal
+82,LawlessEasy
+85,LawlessExpert
+86,LawlessExpertPlus
+84,LawlessHard
+83,LawlessNormal
+92,LegacyEasy
+95,LegacyExpert
+96,LegacyExpertPlus
+94,LegacyHard
+93,LegacyNormal
+87,LightShowEasy
+90,LightShowExpert
+91,LightShowExpertPlus
+89,LightShowHard
+88,LightShowNormal
+77,NoArrowsEasy
+80,NoArrowsExpert
+81,NoArrowsExpertPlus
+79,NoArrowsHard
+78,NoArrowsNormal
+72,OneSaberEasy
+75,OneSaberExpert
+76,OneSaberExpertPlus
+74,OneSaberHard
+73,OneSaberNormal
+67,StandardEasy
+70,StandardExpert
+71,StandardExpertPlus
+69,StandardHard
+68,StandardNormal
