+++ database/data/genre.csv
@@ -1,1 +1,37 @@
-2,Pop
+12,Alternative
+13,Ambient
+14,Anime
+15,Classical-Orchestral
+16,Comedy-Meme
+17,Dance
+18,Drum and Bass
+19,Dubstep
+20,Electronic
+21,Folk-Acoustic
+22,Funk-Disco
+23,Hardcore
+24,Hip Hop-Rap
+25,Holiday
+26,House
+27,Indie
+28,Instrumental
+29,J-Pop
+30,J-Rock
+31,Jazz
+32,K-Pop
+33,Kids-Family
+34,Metal
+35,Nightcore
+11,NULL
+36,Pop
+37,Punk
+38,R&B
+39,Rock
+40,Soul
+41,Speedcore
+42,Swing
+44,Techno
+45,Trance
+43,TV-Film
+46,Video Game
+47,Vocaloid
