--- database/data/note_probabilities.csv
@@ -1,1 +1,0 @@
-5,120,5,3,12,14,1,2,2
