+++ database/data/tag.csv
@@ -1,1 +1,8 @@
-2,Speed
+1,Accuracy
+5,Balanced
+6,Challenge
+7,Dance
+8,Fitness
+3,NULL
+9,Speed
+10,Tech
